+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added the state registration field to the cost center UI
+
 ## [1.15.0] - 2022-10-21
 
 ### Changed